Fixed char representation, Java chars are Unicode codepoints which might not be actual chars

+++ src/annotation/smali.rs
@@ -178,7 +178,7 @@ mod tests {
                             },
                             AnnotationParameter {
                                 name: "charValue".to_string(),
-                                value: AnnotationParameterValue::Literal(Literal::Char('2')),
+                                value: AnnotationParameterValue::Literal(Literal::Char('2' as u16)),
                             },
                             AnnotationParameter {
                                 name: "shortValue".to_string(),
